E2e system tests

+++ hedera-nft-auction-demo-java-node/src/main/java/com/hedera/demo/auction/node/app/db/Sequences.java
@@ -17,5 +17,5 @@ public class Sequences {
     /**
      * The sequence <code>public.auctions_id_seq</code>
      */
-    public static final Sequence<Integer> AUCTIONS_ID_SEQ = Internal.createSequence("auctions_id_seq", Public.PUBLIC, org.jooq.impl.SQLDataType.INTEGER.nullable(false), null, null, null, null, false, null);
+    public static final Sequence<Long> AUCTIONS_ID_SEQ = Internal.createSequence("auctions_id_seq", Public.PUBLIC, org.jooq.impl.SQLDataType.BIGINT.nullable(false), null, null, null, null, false, null);
 }
